No longer a need for a separate addText in the frontend form class
